ConDec-359: Rename test classes
