+++ src/test/java/de/uhd/ifi/se/decision/management/jira/persistence/jiraissuepersistencemanager/TestDeleteDecisionKnowledgeElement.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.persistence.issuestrategy;
+package de.uhd.ifi.se.decision.management.jira.persistence.jiraissuepersistencemanager;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
