+++ src/test/java/de/uhd/ifi/se/decision/management/jira/persistence/jiraissuepersistencemanager/TestInsertDecisionKnowledgeElement.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.persistence.issuestrategy;
+package de.uhd.ifi.se.decision.management.jira.persistence.jiraissuepersistencemanager;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
